refactor : 기존 Modal 제거
