+++ src/app/grouping/create/components/inputSection/MaxUserSection.client.tsx
@@ -2,7 +2,7 @@
 import { useCreateGroupContext } from '../CreateGroupContext';
 import InputArea from '../InputArea.server';
 import { BottomFixedButton } from '@/components/common/Button';
-import { BottomSheet } from '@/components/common/Modal';
+import { BottomSheet } from '@/components/Modal';
 import { NumberSwipePicker } from '@/components/common/SwipePicker';
 import useModalState from '@/store/useModalStore';
