+++ src/app/join/funnels/step4/components/GenderSection.client.tsx
@@ -2,7 +2,7 @@
 import { useJoinContext } from '../../../components/JoinContext';
 import { BottomFixedButton } from '@/components/common/Button';
 import { Input } from '@/components/common/Input';
-import { BottomSheet } from '@/components/common/Modal';
+import { BottomSheet } from '@/components/Modal';
 import GenderSwipePicker from '@/components/common/SwipePicker/GenderSwipePicker';
 import useModalStore from '@/store/useModalStore';
