+++ src/components/Modal/ModalContext.client.tsx
@@ -18,4 +18,5 @@ const useModalContext = () => {
   if (!ctx) throw new Error('Cannot find Context. It should be wrapped within ContextProvider.');
   return ctx;
 };
+
 export { useModalContext };
