+++ src/components/Modal/index.ts
@@ -1,4 +1,4 @@
 export { default as ModalWrapper } from './ModalWrapper.client';
 export { default as BottomSheet } from './BottomSheet';
 export { default as Modal } from './Modal.client';
-export { default as ModalContext } from './ModalContext.client';
+export * from './ModalContext.client';
