--- src/components/common/Modal/index.ts
@@ -1,3 +1,0 @@
-export { default as ModalWrapper } from './ModalWrapper';
-export { default as BottomSheet } from './BottomSheet.client';
-export { default as ConfirmModal } from './ConfirmModal.client';
